corrected tests

+++ src/App.test.js
@@ -25,6 +25,6 @@ test('expect 100 + 20 to equal 120', () =>
 test('expect 4 + 3 to equal 7', () => expect(Numbers.addTwo(4, 3)).toBe(7))
 
 test('expect 2 * 2 to equal 4', () => expect(Numbers.double(2)).toBe(4))
-test('expect 2 * 4 to equal 8', () => expect(Numbers.double(2)).toBe(8))
-test('expect 2 * 100 to equal 200', () => expect(Numbers.double(2)).toBe(200))
-test('expect 2 * 26 to equal 52', () => expect(Numbers.double(2)).toBe(52))
+test('expect 2 * 4 to equal 8', () => expect(Numbers.double(4)).toBe(8))
+test('expect 2 * 100 to equal 200', () => expect(Numbers.double(100)).toBe(200))
+test('expect 2 * 26 to equal 52', () => expect(Numbers.double(26)).toBe(52))
